% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meeting.R
\name{get_meeting}
\alias{get_meeting}
\title{Get speeches by meeting}
\usage{
get_meeting(house = "Lower", sessionNumber = NA, startDate = NA,
  endDate = NA, year = NA, meetingName = NA, ...)
}
\arguments{
\item{house}{Name of the house, value is "Upper", "Lower", or "Both"}

\item{sessionNumber}{integer, session number}

\item{startDate}{starting date to obtain the record in the format of "%Y-%m-%d"
(e.g. "1999-01-01"), if you specified session number, you cannot assign this
option.}

\item{endDate}{ending date to obtain the record in the format of "%Y-%m-%d"
(e.g. "1999-01-01"), if you specified session number, you cannot assign this
option.}

\item{year}{integer, year}

\item{meetingName}{name of the meeting in Japanese. example "予算委員会", "本会議"}

\item{...}{}
}
\value{
the function returns a data.frame of speeches.
}
\description{
This function returns all speeches based on the specified conditions.
Three parameters have to be specified. The first is the name of house
(\code{house}).
The second is the name of meeting
(\code{meetingName}) (e.g. "文教委員会", "本会議").
And the third is the time period. There are three ways to specifiy the time period
(1. starting and ending date, 2. National Diet session number, and 3.
year). If the specified conditions exceed the limit of the number of records
for one API call (2 records per call), this function will repeatedly call the
API until all records are obtained.
}
\examples{
budget_122 <- get_meeting(meetingName = "予算委員会", sessionNumber = 122)
head(budget_122)
}
