% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{session_info}
\alias{session_info}
\title{data on the National Diet sessions}
\format{
An object of class \code{data.frame} with 208 rows and 5 columns.
}
\usage{
session_info
}
\description{
data on the National Diet sessions
}
\keyword{datasets}
