% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_speechdf.R
\name{xml_to_speechdf}
\alias{xml_to_speechdf}
\title{Convert Kaigiroku XML to well-forated data frame}
\usage{
xml_to_speechdf(xml_out)
}
\arguments{
\item{xml_out}{an xml object returned from kokkai kaigiroku API}
}
\value{
data.frame
}
\description{
Convert Kaigiroku XML to well-forated data frame
}
\examples{

}
